Header Experimentation
Separated the header into two flex containers, the other for the mobile search bar and the main one for the majority of the header content. Nearly there with the header, just need to fix an issue with an <a> tag not displaying the way I want it to and then all that's left for the header is the additional navbar.