First Draft Header
Successfully removed the underline and added the hover function to the support link. Also added the correct styles and hover function to the contact link. Corrected the border-radius of the search bar button and added a hover function to it. Experimented with different ideas to try and get the spacing between elements exactly the way it is on the Netmatters website but so far have not been successful. Deleted the backup copies as I am a lot happier with the current files.