Banner Experimentation
Moved onto the main body, I have started with the banner. The source code helps and I am currently going through it trying to understand how to set up my HTML code to have an image with overlaying text with the correct styles applied to it for the relevant breakpoints.

There are 3 main breakpoints that the banner adapts to and one minor breakpoint.

The adaptations seemed to be caused by the header more so than the banner, but there are properties present that allow the banner to adapt at this minor breakpoint.

Image and styles have been added for the banner however there seems to be some issues with getting the image to display and behave the way it does on the Netmatters website. For some reason the object-fit property and cover value isn't having the same effect on my image as it does on the Netmatters website.

Will continue to experiment tomorrow.