Late Re-do of header
I created a new branch to experiment with adding the original source code and CSS styles to see if I could get it to work the same as the Netmatters webpage.

The inheritance issues are sorted however the action links do not display the way I need them to be.

I will most likely go back to the original branch and carry on from there, taking the bits of code from this branch that have been useful.