Realisation About My Code
After talking to more experienced scions and developers on my course it's become clear that I have massively over-complicated the task and the task for myself.

I have removed a lot of the additional divs and stripped it back to the basics, a little more experimentation with the banner and I will have simplified that successfully.

The header in this commit is a much simpler yet just as effective header.