Middle Section
I have sourced and placed all the icons as well as the circular element placed behind them.

All that's left for this middle section is to add the hover feature to the blocks, since this is the last night I have a guaranteed wi-fi connection I will push these commits to github and submit this version as my first draft for my HTML & CSS Reflection task.