Feedback Round 1
Cleaned up the code a bit, added a service-cards.scss and a fonts.scss.

Used a combination of flex-box and grid to get the cards to appear the way they should. All that's left for the fonts is to get the correct placing on the "read more" button and import fonts for them.